rename commands folder to cli
