+++ src/cli/generate.ts
